🏗 [portafly] moves pages folder into components
